+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - InputDefinition::isValidPath to support absolute paths
 - schema docprint utility support of absolute paths
 - MCNP STRING token pattern to allow for h-h2o.42t, etc.
+- InputDefinition to successfully build if ChildCountEqual rule uses lookup value
 
 ### Changed
 - Minimum version of CMake to 3.20, removing deprecated version incompatibility/error observed with newer CMake